added gpm models
